small change
i made it so when you select a game you get an alert saying the game youve selected. we can change this later and adjust it so it can be added to the favorites tab like the cat api. then we work on cart.

also i added 2 more games the the home screen.